LayoutTests/fast/page-color-sampling/color-sampling-with-full-viewport-modal-container-expected.txt

PASS initialColors.top is "rgb(255, 99, 71)"
PASS initialColors.bottom is null
PASS finalColors.top is "rgb(255, 99, 71)"
PASS finalColors.bottom is "rgb(200, 100, 100)"
PASS successfullyParsed is true

TEST COMPLETE
